pyproject.toml
[project]
name = "faststream-monitoring"
version = "0.1.0"
description = "Example of monitoring settings for FastStream."
requires-python = ">=3.12"
readme = "README.md"
authors = [
    {name = "treaditup", email = "treaditup@gmail.com"},
]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "faststream @ git+https://github.com/draincoder/faststream.git@opentelemetry",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "adaptix>=3.0.0b5",
    "aio-pika>=9.4.1",
]

[project.urls]
"Source" = "https://github.com/draincoder/faststream-monitoring"
"Homepage" = "https://github.com/draincoder/faststream-monitoring"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm.dev-dependencies]
lint = [
    "ruff>=0.3.4",
    "mypy>=1.9.0",
    "pre-commit>=3.7.0",
]
test = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
]

[tool.ruff]
line-length = 120
target-version="py312"
src = ["src"]
include = ["src/**.py", "tests/**.py"]

lint.select = [
    "ALL",
]

lint.ignore = [
    "ARG",
    "ANN",
    "D",
    "EM101",
    "EM102",
    "PT001",
    "PT023",
    "SIM108",
    "SIM114",
    "TRY003",
    "PLW2901",
    "RET505",
    "ISC001",
    "UP035",
    "COM812",
    "PTH123",
    "S311",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["TID252", "PLR2004", "S101", "T201", "SLF001", "PERF203", "S603", "S607"]

[tool.ruff.lint.mccabe]
max-complexity = 12

[tool.ruff.lint.isort]
known-first-party = ["src", "tests"]
no-lines-before = ["local-folder"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 88

[tool.mypy]
packages = ["src", "tests"]
python_version = "3.12"

strict = true
pretty = true
allow_redefinition = true
allow_untyped_calls = true
disallow_any_explicit = false
disallow_any_generics = false
disallow_subclassing_any = false
disallow_untyped_decorators = false
ignore_missing_imports = true
show_column_numbers = true
show_error_context = true
warn_unreachable = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = "ignore::DeprecationWarning"
